storage: move route_dispenser to the separate file
Route dispenser is a simple container for routes, let's move it out from
the storage file to the separate one in the scope of #263. While
we're here let's also rewrite its test to luatest in the scope of #371.

Part of #263
Part of #371

NO_DOC=refactoring
NO_TEST=refactoring
