+++ test/unit/rebalancer.result
@@ -16,9 +16,6 @@ build_routes = vshard.storage.internal.rebalancer_build_routes
 calc_etalon = require('vshard.replicaset').calculate_etalon_balance
 ---
 ...
-dispenser = vshard.storage.internal.route_dispenser
----
-...
 rlist = vshard.storage.internal.rlist
 ---
 ...
@@ -1068,161 +1065,6 @@ build_routes(replicasets)
     uuid4: 2500
   uuid2:
     uuid4: 2500
-...
---
--- gh-161: parallel rebalancer. One of the most important part of
--- the latter is a dispenser. It is a structure which hands out
--- destination UUIDs in a round-robin manner to worker fibers.
---
-d = dispenser.create({uuid = 15})
----
-...
-dispenser.pop(d)
----
-- uuid
-...
-for i = 1, 14 do assert(dispenser.pop(d) == 'uuid', i) end
----
-...
-dispenser.pop(d)
----
-- null
-...
-dispenser.pop(d)
----
-- null
-...
-dispenser.pop(d)
----
-- null
-...
-dispenser.pop(d)
----
-- null
-...
-d
----
-- rlist:
-    count: 0
-  map:
-    uuid:
-      bucket_count: 0
-      id: uuid
-      progress: 0
-      is_throttle_warned: false
-      need_to_send: 15
-...
-d = dispenser.create({uuid1 = 5, uuid2 = 5})
----
-...
-u = dispenser.pop(d)
----
-...
-u, d
----
-- uuid2
-- rlist:
-    count: 2
-    last: &0
-      bucket_count: 4
-      prev: &1
-        bucket_count: 5
-        next: *0
-        id: uuid1
-        progress: 0
-        is_throttle_warned: false
-        need_to_send: 5
-      id: uuid2
-      progress: 0
-      is_throttle_warned: false
-      need_to_send: 5
-    first: *1
-  map:
-    uuid1: *1
-    uuid2: *0
-...
-dispenser.put(d, u)
----
-...
-d
----
-- rlist:
-    count: 2
-    last: &0
-      bucket_count: 5
-      prev: &1
-        bucket_count: 5
-        next: *0
-        id: uuid1
-        progress: 0
-        is_throttle_warned: false
-        need_to_send: 5
-      id: uuid2
-      progress: 0
-      is_throttle_warned: false
-      need_to_send: 5
-    first: *1
-  map:
-    uuid1: *1
-    uuid2: *0
-...
-dispenser.throttle(d, u)
----
-- true
-...
-d
----
-- rlist:
-    count: 2
-    last: &0
-      bucket_count: 5
-      prev: &1
-        bucket_count: 5
-        next: *0
-        id: uuid1
-        progress: 0
-        is_throttle_warned: false
-        need_to_send: 5
-      id: uuid2
-      progress: 0
-      is_throttle_warned: true
-      need_to_send: 5
-    first: *1
-  map:
-    uuid1: *1
-    uuid2: *0
-...
-dispenser.throttle(d, u)
----
-- false
-...
-u1 = dispenser.pop(d)
----
-...
-u2 = dispenser.pop(d)
----
-...
-u1, u2
----
-- uuid1
-- uuid2
-...
-for i = 1, 4 do					\
-	assert(dispenser.pop(d) == u1)		\
-	assert(dispenser.pop(d) == u2)		\
-end
----
-...
--- Double skip should be ok. It happens, if there were several
--- workers on one destination, and all of them received an error.
-d = dispenser.create({uuid1 = 1})
----
-...
-dispenser.skip(d, 'uuid1')
----
-...
-dispenser.skip(d, 'uuid1')
----
 ...
 _bucket:drop()
 ---
