+++ vshard/storage/CMakeLists.txt
@@ -1,3 +1,3 @@
 install(FILES init.lua reload_evolution.lua ref.lua sched.lua schema.lua
-        export_log.lua exports.lua
+        export_log.lua exports.lua route_dispenser.lua
         DESTINATION ${TARANTOOL_INSTALL_LUADIR}/vshard/storage)
